Extract DCR config creation to DotcomRenderingConfig
