+++ sport/app/football/model/DotcomRenderingCricketDataModel.scala
@@ -3,16 +3,13 @@ package football.model
 import common.{CanonicalLink, Edition}
 import conf.Configuration
 import cricket.controllers.CricketMatchPage
-import cricketModel.{Innings, InningsBatter, InningsBowler, InningsWicket, Match, Team}
-import experiments.ActiveExperiments
+import cricketModel.{Innings, InningsWicket, Match, Team}
 import model.ApplicationContext
-import model.dotcomrendering.DotcomRenderingUtils.{assetURL, withoutNull}
-import model.dotcomrendering.{Config, PageFooter, PageType}
+import model.dotcomrendering.DotcomRenderingUtils.withoutNull
+import model.dotcomrendering.{DotcomRenderingConfig, PageFooter, PageType}
 import navigation.{FooterLinks, Nav}
 import play.api.libs.json.{JsObject, JsValue, Json, Writes}
 import play.api.mvc.RequestHeader
-import views.support.{CamelCase, JavaScriptPage}
-import ab.ABTests
 
 case class DotcomRenderingCricketDataModel(
     cricketMatch: Match,
@@ -40,27 +37,11 @@ object DotcomRenderingCricketDataModel {
 
     val pageType: PageType = PageType(page, request, context)
 
-    val switches: Map[String, Boolean] = conf.switches.Switches.all
-      .filter(_.exposeClientSide)
-      .foldLeft(Map.empty[String, Boolean])((acc, switch) => {
-        acc + (CamelCase.fromHyphenated(switch.name) -> switch.isSwitchedOn)
-      })
-
-    val config = Config(
-      switches = switches,
-      abTests = ActiveExperiments.getJsMap(request),
-      serverSideABTests = ABTests.getParticipations,
-      ampIframeUrl = assetURL("data/vendor/amp-iframe.html"),
-      googletagUrl = Configuration.googletag.jsLocation,
-      stage = common.Environment.stage,
-      frontendAssetsFullURL = Configuration.assets.fullURL(common.Environment.stage),
+    val combinedConfig = DotcomRenderingConfig(
+      page = page,
+      request = request,
+      isPreview = pageType.isPreview,
     )
-
-    val combinedConfig: JsObject = {
-      val jsPageConfig: Map[String, JsValue] =
-        JavaScriptPage.getMap(page, Edition(request), pageType.isPreview, request)
-      Json.toJsObject(config).deepMerge(JsObject(jsPageConfig))
-    }
 
     DotcomRenderingCricketDataModel(
       page.theMatch,
